+++ CHANGELOG.md
@@ -88,6 +88,8 @@
 
 - Fix networking indicator progress bar position on the screen (INDIGO Sprint 220204, [!234](https://github.com/TeskaLabs/asab-webui/pull/234))
 
+- Move Library module from lmio-webui to asab-webui (INDIGO Sprint 220204, [!231](https://github.com/TeskaLabs/asab-webui/pull/231))
+
 ### Refactoring
 
 - Renaming configuration option FAKE_USERINFO to MOCK_USERINFO and refactoring code accordingly (INDIGO Sprint 210406, [!91](https://github.com/TeskaLabs/asab-webui/pull/91))